fix ios breakage

+++ src/entities/route_plan.rs
@@ -3,6 +3,6 @@ use crate::entities::*;
 #[derive(Debug)]
 pub struct RoutePlan {
     pub route_points: Vec<GeoPoint>,
-    pub min_distance: u32,
-    pub min_points: u32,
+    pub min_distance: u64,
+    pub min_points: u64,
 }
