+++ src/entities/user.rs
@@ -3,7 +3,7 @@ pub struct User {
     pub username: String,
     pub password: String,
     pub campus_name: String,
-    pub uid: u32,
-    pub unid: u32,
+    pub uid: u64,
+    pub unid: u64,
     pub token: String,
 }
