+++ src/utils.rs
@@ -3,8 +3,8 @@ use rand::{self, Rng};
 use itertools::Itertools;
 use md5;
 
-pub fn rand_near(base: u32, err: u32) -> u32 {
-    (base as f64 + (err as f64 * (rand::thread_rng().next_f64() * 2.0 - 1.0))) as u32
+pub fn rand_near(base: u64, err: u64) -> u64 {
+    (base as f64 + (err as f64 * (rand::thread_rng().next_f64() * 2.0 - 1.0))) as u64
 }
 
 pub fn rand_near_f64(base: f64, err: f64) -> f64 {
